requirements.txt
streamlit==1.4.0
pandas-datareader==0.10.0
numpy==1.20.3
matplotlib==3.1.3